--- pch/janus_ftl_test_pch.h
@@ -1,12 +1,0 @@
-/**
- * @file janus_ftl_test_pch.h
- * @author Hayden McAfee (hayden@outlook.com)
- * @date 2021-03-18
- * @copyright Copyright (c) 2021 Hayden McAfee
- */
-
-#include <fmt/core.h>
-#include <spdlog/spdlog.h>
-
-#define CATCH_CONFIG_ALL_PARTS
-#include <catch2/catch.hpp>
