+++ test/test.cpp
@@ -5,16 +5,5 @@
  * @copyright Copyright (c) 2020 Hayden McAfee
  */
 
-// Some Catch2 defines required for PCH support
-// https://github.com/catchorg/Catch2/blob/v2.x/docs/ci-and-misc.md#precompiled-headers-pchs
-#undef TWOBLUECUBES_SINGLE_INCLUDE_CATCH_HPP_INCLUDED
-#define CATCH_CONFIG_IMPL_ONLY
-#define CATCH_CONFIG_RUNNER
+#define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
-
-int main(int argc, char* argv[])
-{
-    // Test!
-    int result = Catch::Session().run(argc, argv);
-    return result;
-}
